ts/src/media_playlist.ts
import { Duration, MaybeTimeDuration } from 'typed-duration'

// Define the MediaSegment interface
export interface MediaSegment {
  duration: MaybeTimeDuration;
  url: string;
}

// Define the MediaPlaylist class
export class MediaPlaylist {
  ended: boolean;
  segments: MediaSegment[];
  target_duration: MaybeTimeDuration;
  version: number;

  constructor() {
    this.ended = false;
    this.segments = [];
    this.target_duration = Duration.seconds.from(0); // Initialize with the appropriate value
    this.version = 0;
  }

  // Define the parseExtM3U method
  static parseExtM3U(_file: string): MediaPlaylist {
    // You can implement the parsing logic here
    return new MediaPlaylist();
  }
}

